pyproject.toml
[tool.poetry]
name = "restful-svelte"
version = "0.1.0"
description = ""
authors = ["Daniel Terreros <110783387+DanielTKC@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
whitenoise = "^6.6.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
